+++ Wire-iOS Tests/Utils/XCTestCase+ImageFromBundle.swift
@@ -46,3 +46,22 @@ extension XCTestCase {
         return image(inTestBundleNamed: "unsplash_matterhorn.jpg").jpegData(compressionQuality: 0.9)!
     }
 }
+
+extension UIImage {
+    public convenience init?(inTestBundleNamed name: String,
+                             for aClass: AnyClass) {
+
+        let bundle = Bundle(for: aClass)
+
+        let url = bundle.url(forResource: name, withExtension: "")
+
+        if let isFileURL = url?.isFileURL {
+            XCTAssert(isFileURL)
+        } else {
+            XCTFail("\(name) does not exist")
+        }
+
+        self.init(contentsOfFile: url!.path)!
+    }
+
+}
