+++ Wire-iOS/Sources/UserInterface/ConversationList/ConversationListViewController+TopBar.swift
@@ -18,7 +18,6 @@
 
 
 import UIKit
-import Cartography
 
 extension ConversationListViewController: NetworkStatusBarDelegate {
     var bottomMargin: CGFloat {
